Move cache to cache hierarchy
