+++ src/main/java/org/apache/cassandra/sidecar/metrics/CacheMetrics.java
@@ -18,116 +18,19 @@
 
 package org.apache.cassandra.sidecar.metrics;
 
-import java.util.Objects;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.LongAdder;
-
-import com.codahale.metrics.Histogram;
 import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.Timer;
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.stats.CacheStats;
-import com.github.benmanes.caffeine.cache.stats.StatsCounter;
-import org.checkerframework.checker.index.qual.NonNegative;
-import org.checkerframework.checker.nullness.qual.NonNull;
 
-import static org.apache.cassandra.sidecar.metrics.SidecarMetrics.APP_PREFIX;
+import static org.apache.cassandra.sidecar.routes.snapshots.ListSnapshotHandler.SNAPSHOT_CACHE_NAME;
 
 /**
- * A {@link StatsCounter} implementation that accumulates statistics during the operation
- * of a {@link Cache}
+ * Tracks metrics related to internal caches.
  */
-public class CacheMetrics implements StatsCounter
+public class CacheMetrics
 {
-    public static final String CACHE_PREFIX = APP_PREFIX + ".Cache";
-    protected final MetricRegistry metricRegistry;
+    public final CacheStatsCounter snapshotCacheMetrics;
 
-    protected final NamedMetric<DeltaGauge> hits;
-    protected final NamedMetric<DeltaGauge> misses;
-    protected final NamedMetric<Timer> loadSuccess;
-    protected final NamedMetric<Timer> loadFailure;
-    protected final NamedMetric<Histogram> evictions;
-    protected final LongAdder totalLoadTimeNanos = new LongAdder();
-
-    public CacheMetrics(MetricRegistry metricRegistry, String cacheName)
+    public CacheMetrics(MetricRegistry globalMetricRegistry)
     {
-        this.metricRegistry = Objects.requireNonNull(metricRegistry, "Metric registry can not be null");
-
-        String domain = CACHE_PREFIX + "." + cacheName;
-        hits = NamedMetric.builder(name -> metricRegistry.gauge(name, DeltaGauge::new))
-                          .withDomain(domain)
-                          .withName("Hits")
-                          .build();
-        misses = NamedMetric.builder(name -> metricRegistry.gauge(name, DeltaGauge::new))
-                            .withDomain(domain)
-                            .withName("Misses")
-                            .build();
-        loadSuccess = NamedMetric.builder(metricRegistry::timer)
-                                 .withDomain(domain)
-                                 .withName("LoadSuccess")
-                                 .build();
-        loadFailure = NamedMetric.builder(metricRegistry::timer)
-                                 .withDomain(domain)
-                                 .withName("LoadFailure")
-                                 .build();
-        evictions = NamedMetric.builder(metricRegistry::histogram)
-                               .withDomain(domain)
-                               .withName("Evictions")
-                               .build();
-    }
-
-    @Override
-    public void recordHits(@NonNegative int count)
-    {
-        hits.metric.update(count);
-    }
-
-    @Override
-    public void recordMisses(@NonNegative int count)
-    {
-        misses.metric.update(count);
-    }
-
-    @Override
-    public void recordLoadSuccess(@NonNegative long loadTime)
-    {
-        loadSuccess.metric.update(loadTime, TimeUnit.NANOSECONDS);
-        totalLoadTimeNanos.add(loadTime);
-    }
-
-    @Override
-    public void recordLoadFailure(@NonNegative long loadTime)
-    {
-        loadFailure.metric.update(loadTime, TimeUnit.NANOSECONDS);
-        totalLoadTimeNanos.add(loadTime);
-    }
-
-    @Override
-    public void recordEviction()
-    {
-        recordEviction(1);
-    }
-
-    /**
-     * @deprecated
-     */
-    @Deprecated
-    public void recordEviction(int weight)
-    {
-        evictions.metric.update(weight);
-    }
-
-    @Override
-    public @NonNull CacheStats snapshot()
-    {
-        return CacheStats.of(hits.metric.getValue(), misses.metric.getValue(),
-                             loadSuccess.metric.getCount(), loadFailure.metric.getCount(),
-                             totalLoadTimeNanos.sum(), evictions.metric.getCount(), 0);
-    }
-
-    @Override
-    public String toString()
-    {
-        return snapshot().toString();
+        snapshotCacheMetrics = new CacheStatsCounter(globalMetricRegistry, SNAPSHOT_CACHE_NAME);
     }
 }
